IntervalArray equality follow-ups
